fix percentile

+++ integ-test/src/test/java/org/opensearch/sql/calcite/standalone/CalcitePPLAggregationIT.java
@@ -608,7 +608,7 @@ public class CalcitePPLAggregationIT extends CalcitePPLIntegTestCase {
             String.format(
                 "source=%s | stats percentile(balance, 50) as p50, percentile(balance, 90) as p90",
                 TEST_INDEX_BANK));
-    verifySchema(actual, schema("p50", "double"), schema("p90", "double"));
+    verifySchema(actual, schema("p50", "long"), schema("p90", "long"));
     verifyDataRows(actual, rows(32838, 48086));
   }
